fix: ab detection functionality not working in the build (#3125)
* ab detection debug tool : plugin added

* ab detection tracker: work in progress

* ab-gltf detection debug tool initial implementation finished

* post review fixes

* ab detection asset and tag fix

* detectabs for single scene debugged

* ab detector tracker algo bug fix

* ab detection additional fixes

* move detectabs method from main to bridges

* bugfix: ab detector tracking stopped turning off if you call turning it on twice

* bugfix: ab detection stopped turning off for certain order of turn on commands
