+++ unity-renderer/Assets/Prefabs/Resources/Bridges.prefab
@@ -19,6 +19,7 @@ GameObject:
   - component: {fileID: 6402082374581292810}
   - component: {fileID: 7142775002132873306}
   - component: {fileID: 607285373800589754}
+  - component: {fileID: 2728691134734170003}
   m_Layer: 0
   m_Name: Bridges
   m_TagString: Untagged
@@ -158,5 +159,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 380981e8faeb448f8dc7a15639ef91af, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &2728691134734170003
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6401575379531892522}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3b06bacb2e6c4c9ba93a35a680d139ed, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
