+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Bridges/AbDetectorBridge/AbDetectorBridge.asmdef
@@ -1,0 +1,4 @@
+﻿{
+  "name": "AbDetectorBridge",
+  "references":[ "GUID:28f74c468a54948bfa9f625c5d428f56", "GUID:1e6b57fe78f7b724e9567f29f6a40c2c" ]
+}
