+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PluginSystem/PluginSystemFactory.cs
@@ -51,7 +51,6 @@ namespace DCL
                 NotificationScriptableObjects.newApprovedFriends,
                 DataStore.i));
             
-            //TODO ANTON make feature flag if needed
             pluginSystem.Register<ABDetectorPlugin>(() => new ABDetectorPlugin());
 
             pluginSystem.SetFeatureFlagsData(DataStore.i.featureFlags.flags);
