+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/RendereableAssetLoadHelper/RendereableAssetLoadHelper.cs
@@ -15,6 +15,7 @@ namespace DCL.Components
         }
         
         private const string FROM_ASSET_BUNDLE_TAG = "FromAssetBundle";
+        private const string FROM_RAW_GLTF_TAG = "FromRawGLTF";
 
         public static bool VERBOSE = false;
 
@@ -228,6 +229,11 @@ namespace DCL.Components
                     meshDataSize = x.meshDataSize,
                     animationClips = x.animationClips
                 };
+
+                foreach (var someRenderer in r.renderers)
+                {
+                    someRenderer.tag = FROM_RAW_GLTF_TAG;
+                }
 
                 OnSuccessWrapper(r, OnSuccess);
             };
